Shorten temp ZIP path length, Move retroglue
