+++ libretro/retroglue.c
